풀 리퀘스트 병합 #59
* 상품조회 Mapping_HomeController로 이동함

* Merge branch 'main' of https://github.com/zziyeon/GREAT into 김지연1007

* 상품조회 Mapping_HomeController로 이동함

* 관리자 헤더 추가

* enenen

* Merge branch 'main' into 김지연1007

* Merge remote-tracking branch 'origin/김지연1007' into 김지연1007

* Merge branch 'main' of https://github.com/zziyeon/GREAT into 김지연1007

* enenen

* enenen

+++ src/main/resources/static/js/product/manage.js
@@ -161,13 +161,11 @@ function each_sell_status(obj, pNumber, pstatus){
         sell_st_uf(pNumber, obj.value);
     }
     if(obj.value=='del'){
-//        if(confirm("판매글을 삭제하시겠습니까? \n삭제 후 복구 불가합니다."))
-//            window.location.href = "/products/" + pNumber + "/del";
-//        else{
-//            obj.value=pstatus;
-//        }
-        deleteProduct(pNumber);
-
+        if(confirm("판매글을 삭제하시겠습니까? \n삭제 후 복구 불가합니다."))
+            deleteProduct(pNumber);
+        else{
+            obj.value=pstatus;
+        }
     }
 
 }
